Creating repository with old files
This in order to get all the benefits from git, will add to gitHub asap.

This project is created and executed with the intent to learn C. It will
contain errors and bad coding habits. These will (hopefully) be
corrected in future commits.

As of this first commit, there exists a framework (built upon the
PDCurses library) for a roguelike, with a traversable map, a small
selection of monsters, a barebones combat system and a leveling system.